+++ include/hyper/motion/interpolators/spatial/forward.hpp
@@ -4,7 +4,7 @@
 #pragma once
 
 #include "hyper/definitions.hpp"
-#include "hyper/state/interpolators/temporal/forward.hpp"
+#include "hyper/motion/interpolators/temporal/forward.hpp"
 
 namespace hyper {
