+++ include/hyper/motion/interpolators/temporal/forward.hpp
