+++ include/hyper/motion/interpolators/temporal/temporal.hpp
@@ -3,7 +3,7 @@
 
 #pragma once
 
-#include "hyper/state/interpolators/temporal/forward.hpp"
+#include "hyper/motion/interpolators/temporal/forward.hpp"
 
 namespace hyper {
