+++ include/hyper/motion/motion.hpp
@@ -3,7 +3,7 @@
 
 #pragma once
 
-#include "hyper/state/forward.hpp"
+#include "hyper/motion/forward.hpp"
 
 namespace hyper {
