+++ internal/hyper/motion/continuous.cpp
@@ -4,7 +4,7 @@
 #include <glog/logging.h>
 
 #include "hyper/matrix.hpp"
-#include "hyper/state/continuous.hpp"
+#include "hyper/motion/continuous.hpp"
 #include "hyper/variables/groups/se3.hpp"
 #include "hyper/variables/stamped.hpp"
